+++ code/src/sixsq/nuvla/server/resources/group.clj
@@ -38,8 +38,7 @@ that start with 'nuvla-' are reserved for the server.
 
 (def collection-acl {:query       ["group/nuvla-user"]
                      :add         ["group/nuvla-admin"
-                                   "group/nuvla-user"]
-                     :bulk-delete ["group/nuvla-admin"]})
+                                   "group/nuvla-user"]})
 
 
 ;;
@@ -250,14 +249,6 @@ that start with 'nuvla-' are reserved for the server.
 (defmethod crud/query resource-type
   [request]
   (query-impl request))
-
-
-(def bulk-delete-impl (std-crud/bulk-delete-fn resource-type collection-acl collection-type))
-
-
-(defmethod crud/bulk-delete resource-type
-  [request]
-  (bulk-delete-impl request))
 
 
 ;;