+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 ## Unreleased
+### Added
+
+- Group - Group able to create a subgroup
+
+### Changed
+
+- Group - Bulk delete removed not needed
 
 ## [5.25.2] - 2022-05-12